LayoutTests/platform/ios/imported/w3c/web-platform-tests/html/canvas/offscreen/text/2d.text.measure.emHeights-expected.txt

2d.text.measure.emHeights

Testing emHeights for OffscreenCanvas


FAIL Testing emHeights for OffscreenCanvas assert_equals: ctx.measureText('A').emHeightAscent === 37.5 (got 86[number], expected 37.5[number]) expected 37.5 but got 86
